monitoring/requirements.txt
# FlashFit AI Monitoring System Requirements
# Pinned versions for stability and reproducibility

# Core Monitoring
prometheus_client==0.17.1
psutil==5.9.5
requests==2.32.0

# Data Processing
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.1
seaborn==0.12.3

# Configuration / Environment
pyyaml==6.0
python-dotenv==1.0.0

# Async / File Handling
aiohttp==3.9.5
aiofiles==23.2.1

# Database
psycopg2-binary==2.9.9
redis==5.2.0

# Logging
structlog==24.2.0
rich==13.6.0